+++ conftest.py
@@ -1,14 +1,18 @@
 #
 # Common settings and fixtures used with pytest
 #
+import datetime
+
 import pytest
 from django.core.management import call_command
 import logging
 
+from freezegun import freeze_time
 from trackstats.models import Domain, Metric
 
 from topobank.analysis.registry import AnalysisFunctionRegistry
-from topobank.manager.tests.utils import UserFactory
+from topobank.manager.tests.utils import UserFactory, SurfaceFactory
+from topobank.users.tests.factories import UserFactory
 
 _log = logging.getLogger(__name__)
 
@@ -95,3 +99,22 @@ def reset_singletons():
     del reg
 
 
+@pytest.mark.django_db
+@pytest.fixture
+def example_pub():
+    """Fixture returning a publication which can be used as test example"""
+
+    user = UserFactory()
+
+    authors = "Alice, Bob"
+    publication_date = datetime.date(2020, 1, 1)
+    description = "This is a nice surface for testing."
+    name = "Diamond Structure"
+
+    surface = SurfaceFactory(name=name, creator=user, description=description)
+    surface.tags = ['diamond']
+
+    with freeze_time(publication_date):
+        pub = surface.publish('cc0-1.0', authors)
+
+    return pub